Ensure 'BuildOption()' is actually empty
and doesn't override anything.

BuildOptions and the associated case classes shouldn't have Some(…)
values. These make 'BuildOptions().orElse(other)' override values in
'other', which breaks its semantics / assumptions. ('BuildOptions' is
supposed to be the "zero" or null value of BuildOptions as a monoid, and
is used more as such with the changes in this PR.)