Make bloop start test more reliable
When using a native Scala CLI launcher, -Xmx1k was used by scala-cli
itself, and was making Scala CLI rather than Bloop crash. With earlier
changes in this PR, it seems the 1k is enough for scala-cli to proceed.
It then complains than --bloop-java-opt is missing an argument (as
-Xmx1k was taken by some native-image stuff for itself).

Passing an argument that doesn't look like a Java option just works
instead.