Start BSP with sensible default target id
We're trying to guess the target id that Bloop is going to give us.
Depending on whether we use Java IO or NIO, or if the .scala directory
already exists, the URI we infer might be subtly different than the one
picked by Bloop, so that's a bit fragile…

This URI used to be picked in the first messages we got from Bloop.

It's useful here to reset directives diagnostics upon startup.