fix: stick to a tcp connection for Bloop on FreeBSD.
I'll be honest, I haven't really dug into the JNI/JNA issues, so I'm
unsure if there is a better more appropriate to get domain sockets to
work on FreeBSD in this scenario, but when testing it with --bloop-bsp-protocol
being set to tcp it worked. So this is just a quick fix to ensure that the
default for FreeBSD for now is tcp.

Closes #271