Directory layout
My experience was that if "HelloWorld.scala" and "MyTests.test.scala" were not in the canonical src / test directories, no run or test annotations appeared. 

I assume scala-cli doesn't set out to loosen that restriction? If so then please ignore this. 

For someone completely "new" to scala, this requirement may be non-obvious and frustrating, hence the PR. The intent is for the screenshot to make the locations of these files explicit. my apologies that it doesn't appear to follow the <img src={} /> layout - I wasn't able to understand how to contribute that.