Don't put VCS details in Constants.scala when not needed
This is useful when developing on any of the stubs modules (stubs,
runner, test-runner), but this triggers a recompilation of build very
often. This is now disabled by default. When developing on the stubs
modules, set `local-repo`.developingOnStubModules to true in build.sc,
for Mill's watch mode to take into account changes in them.