Move all cli *Options classes to dedicated module
We need this to compile only that module with 2.13 and rest with 3.x

This also mean that I needed to move a lot of logic from *Options classes.
Most such methods was moved to places where it was used.
Rest was moved to dedicated implicit classes, that are later added to `commands` package using various traits from `util` package