In Doctor, account for absolute and relative paths
When checking PATH, some of the entries there are relative.
According to os-lib documentation, Path will throw
exception if you try to use it to construct a relative
path. RelPath needs to be used in this case. This happens,
e.g. when the path starts with `~` which is not handled
by the previous filter in the code.