Unnecessary?
This computes BuildOptions.projectParams early, which itself relies on
BuildOptions.scalaParams, which downloads the supported Scala version
listing. For pure Java projects, we'd rather not have this file be
downloaded.

Later on, BuildOptions are tweaked, so that BuildOptions.scalaParams
doesn't try to download a version listing, for pure Java projects.