Make redirection of Bloop output to file more robust
Previously, we were calling redirectOutput / redirectError to redirect
the Bloop output to a (single) file. But the resulting output file
seemed to contain truncated lines, or to be missing some output. Same
with redirectErrorStream(true). My guess is that the JVM was opening the
output file twice (once for stdout, once for stderr), and writing to the
two file descriptors at the same time was a problem.

With these changes, we start the Bloop server via a shell script, and
make it do "> output-file 2>&1". The shell handles redirecting both
stdout and stderr to the same file better it seems, opening the file
only once under-the-hood I think.

Rather than doing this via a shell script, we could try to do the same
as what the shell does, by calling ourselves system calls to open the
output file (open), duplicating it (dup2?), then fork / exec. That's
more cumbersome (we'd likely have to rely on JNI, or JNA + GraalVM API,
to have it work both from the JVM and the native launcher), so we should
probably do it only if ever using the shell ends up being a problem IMO…