Stop compiling most stuff with Scala 2
It seems both the Scala 2 and Scala 3 versions of core, options, and
all, end up in the cli module class path. With newer developments, I'm
getting NoSuchMethod exceptions at runtime because of this.

At first, we could keep cli2, but we'd need to have build-macros for
Scala 3 expose Scala 2 macros too. It would be possible with simple
standard looking macros, but here, EitherCps.Helper.apply is a macro in
Scala 2, but a standard method in Scala 3, we can't package both in the
same JAR.