chore(ci): add in mill-github-dependency-graph
So no pressure ha, but I'm wondering if you'd be willing to be a guinea
pig for this. @adpi2 at the Scala Center has been working on
https://github.com/scalacenter/sbt-dependency-graph-action which submits
your sbt projects dependency graph to GitHub to be able to both view
your dependencies in the insights tab on GitHub and also to get
dependabot alerts about them if you'd like. I wanted to also offer this
for mill, which is what you're seeing here. You can view the project
here https://github.com/ckipp01/mill-github-dependency-graph, but since
there isn't a lot of larger Mill projects out there I'd love to test it
on some more before announcing it. Locally this works fine with
scala-cli, but it'd be great to actually see it submit.

Any interest in being a tester? :D The only thing you'd need to do is
ensure that the Dependency Graph feature is enabled, and then optionally
the dependabot alerts if you want them. These settings are under
`Settings -> Code security and analysis.`

If this isn't something you want in the repo, no worries at all!