Run commands with a manifest JAR if needed
On Windows, command line length cannot be over some limit. A typical
way to overcome it to run "java -cp" commands is to use a so called
"manifest JAR", that contains the class path in META-INF/MANIFEST.MF
(sole file in the manifest JAR).

Note that some tools don't support manifest JARs (Ammonite for
example, although that could probably be fixed in it).