feat: use the new ScalaAction from BSP4J
This is a simplified version of
https://github.com/VirtusLab/scala-cli/pull/1969 that gets rid of the
`ActionableDiagnostic` that is used internally and instead just ensures
that any time a `TextEdit` is created there is a `title`. Then we take
all the information already contained in the `Diagnostic` and make a
`ScalaAction` from it to be forwarded via the diagnostic data.