docs: remove `.` from command snippet
The impression of `by running:\n\t${SOME COMMAND}` is that you're displaying a console command with the expectation that users may copy paste that line. But this includes a period at the end of the command, weirdly mixing plain english docs punctuation with where a literal command should be. Users who copy paste the line will get the error
```
$ scala-cli config power true.
[error]  Config DB error: Malformed value 'true.' for the 'power' entry, expected a single boolean value ('true' or 'false').
```
until they realize they need to omit the `.`