improvement: Use distinct on ScalacOpt
I am still getting an issue with multiple same plugin options and my guess is that the distinct doesn't work, especially that it is hard to figure out what can actually be contained there. Moving distinct to the end should be a much safer option.

Related to https://github.com/VirtusLab/scala-cli/issues/2708