improvement: Detect objects with main class in scripts
Prebiously, if user had a legacy script with main method then it would not be picked up at all. Now, when we detect the correct signature we try to run it.

The possibility of false positives is pretty low, since user would have to have their own String or Array types.