improvement: Detect objects with main class in scripts (#3479)
* improvement: Detect objects with main class in scripts

Prebiously, if user had a legacy script with main method then it would not be picked up at all. Now, when we detect the correct signature we try to run it.

This will work in case of `def main..` and when object extends App

The possibility of false positives is pretty low, since user would have to have their own App, String or Array types. We will also only use that object if there are no toplevel statements

* improvement: Print both warnings for scripts with main method defined